backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-dotenv
razorpay
python-jose[cryptography]
pydantic
google-generativeai
